+++ explorer/src/lib/components/account-overview/AccountOverview.svelte
@@ -1,14 +1,22 @@
 <script>
-  import { calculateAdaptiveCharCount, middleEllipsis } from "$lib/dusk/string";
-  import { addressCharPropertiesDefaults } from "$lib/constants";
-
-  import "./AccountOverview.css";
   import { onMount } from "svelte";
 
+  import { calculateAdaptiveCharCount, middleEllipsis } from "$lib/dusk/string";
+  import { addressCharPropertiesDefaults } from "$lib/constants";
+  import { OverviewPanelDetailsItem } from "$lib/components";
+  import { luxToDusk } from "$lib/dusk/currency";
+
+  import "./AccountOverview.css";
+  import { createValueFormatter } from "$lib/dusk/value";
+
   export let accountAddress;
+  export let errorFetchingAccountStatus;
+  export let accountBalance;
 
   /** @type {number} */
   let screenWidth = window.innerWidth;
+
+  const formatter = createValueFormatter("en");
 
   onMount(() => {
     const resizeObserver = new ResizeObserver((entries) => {
@@ -42,4 +50,15 @@
       )}
     </b>
   </p>
+  <hr class="account-overview-panel__separator" />
+  <div class="account-overview-panel__details">
+    <OverviewPanelDetailsItem
+      title={errorFetchingAccountStatus
+        ? "N/A"
+        : accountBalance
+          ? `${formatter(luxToDusk(accountBalance))} DUSK`
+          : "Loading..."}
+      subtitle="Current Balance"
+    />
+  </div>
 </section>