+++ explorer/src/lib/mock-data/AccountStatus.d.ts
@@ -1,0 +1,4 @@
+type AccountStatus = {
+  balance: bigint;
+  nonce: bigint;
+};