+++ explorer/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 
 - Add Tokens page [#3415]
+- Add current balance (Account page) [#3564]
 
 ### Changed
 
@@ -166,6 +167,7 @@ and this project adheres to
 [#3415]: https://github.com/dusk-network/rusk/issues/3415
 [#3454]: https://github.com/dusk-network/rusk/issues/3454
 [#3545]: https://github.com/dusk-network/rusk/issues/3454
+[#3564]: https://github.com/dusk-network/rusk/issues/3564
 
 <!-- VERSIONS -->